refactor: move config.py into src/ with root re-export shim
