+++ config.py
@@ -1,212 +1,4 @@
-"""
-Конфигурация приложения
-"""
+"""Backward-compatible re-export. Real module lives in src/config.py."""
+from src.config import Settings, OpenAIModelPreset, settings
 
-import os
-from pydantic_settings import BaseSettings, SettingsConfigDict
-from pydantic import Field, validator, field_validator, BaseModel
-from typing import Optional, List
-
-# Предотвращаем конфликты при форкинге процессов после использования токенизаторов
-os.environ["TOKENIZERS_PARALLELISM"] = "false"
-
-
-class OpenAIModelPreset(BaseModel):
-    key: str = Field(..., description="Уникальный ключ пресета (slug)")
-    name: str = Field(..., description="Отображаемое имя в меню")
-    model: str = Field(..., description="ID модели для API")
-    base_url: Optional[str] = Field(None, description="Базовый URL для этого пресета")
-
-
-class Settings(BaseSettings):
-    """Настройки приложения"""
-    
-    # Telegram Bot
-    telegram_token: str = Field(..., description="Токен Telegram бота")
-    
-    # OpenAI
-    openai_api_key: Optional[str] = Field(None, description="API ключ OpenAI")
-    openai_base_url: Optional[str] = Field(None, description="Базовый URL для OpenAI API")
-    openai_model: str = Field("gpt-3.5-turbo", description="Модель OpenAI для генерации")
-    
-    # Наборы моделей OpenAI с собственными базовыми URL (по одному API ключу)
-    # Формат переменной окружения OPENAI_MODELS: JSON-массив объектов вида
-    # [
-    #   {"key": "openai_gpt4o", "name": "OpenAI: gpt-4o", "model": "gpt-4o", "base_url": "https://api.openai.com/v1"},
-    #   {"key": "local_oq", "name": "OpenRouter: llama3.1:70b", "model": "meta-llama/llama-3.1-70b-instruct", "base_url": "https://openrouter.ai/api/v1"}
-    # ]
-    # Если не задано, автоматически формируется 1 пресет из OPENAI_MODEL и OPENAI_BASE_URL
-    
-    openai_models: List[OpenAIModelPreset] = Field(default_factory=list, description="Список пресетов моделей OpenAI")
-    
-    # Anthropic Claude
-    anthropic_api_key: Optional[str] = Field(None, description="API ключ Anthropic")
-    
-    # Yandex GPT
-    yandex_api_key: Optional[str] = Field(None, description="API ключ Yandex GPT")
-    yandex_folder_id: Optional[str] = Field(None, description="ID папки Yandex Cloud")
-    
-    # LLM Таймауты
-    llm_timeout_seconds: float = Field(30.0, description="Общий таймаут ожидания ответа от LLM (в секундах)")
-    
-    # HTTP заголовки для LLM запросов
-    http_referer: Optional[str] = Field("https://github.com/gihar/Soroka", description="HTTP Referer заголовок для LLM запросов")
-    x_title: Optional[str] = Field("Soroka", description="X-Title заголовок для LLM запросов")
-    
-    # База данных
-    database_url: str = Field("sqlite:///bot.db", description="URL базы данных")
-    
-    # Настройки файлов
-    max_file_size: int = Field(20 * 1024 * 1024, description="Максимальный размер файла в байтах")
-    telegram_max_file_size: int = Field(20 * 1024 * 1024, description="Максимальный размер файла для Telegram Bot API в байтах")
-    max_external_file_size: int = Field(50 * 1024 * 1024, description="Максимальный размер файла из внешних источников (Google Drive, Яндекс.Диск) в байтах")
-    oom_max_file_size_mb: Optional[float] = Field(
-        None,
-        description="Максимальный размер файла для стадии транскрипции (MB). По умолчанию берётся из MAX_FILE_SIZE"
-    )
-    oom_max_memory_percent: float = Field(
-        90.0,
-        description="Максимальный процент использования памяти перед срабатыванием защиты (OOM Killer)"
-    )
-    temp_dir: str = Field("temp", description="Директория для временных файлов")
-    
-    # Логирование
-    log_level: str = Field("INFO", description="Уровень логирования")
-    
-    # SSL настройки
-    ssl_verify: bool = Field(False, description="Проверка SSL сертификатов")
-    
-    # Транскрипция
-    transcription_mode: str = Field("local", description="Режим транскрипции: local (локально), cloud (облако), hybrid (гибридный), speechmatics, deepgram или leopard")
-    groq_api_key: Optional[str] = Field(None, description="API ключ Groq для облачной транскрипции")
-    groq_model: str = Field("whisper-large-v3-turbo", description="Модель Groq для транскрипции")
-    
-    # Speechmatics
-    speechmatics_api_key: Optional[str] = Field(None, description="API ключ Speechmatics для транскрипции и диаризации")
-    speechmatics_language: str = Field("ru", description="Язык для Speechmatics API")
-    speechmatics_domain: Optional[str] = Field(None, description="Домен для Speechmatics (finance, medical и т.д.)")
-    speechmatics_operating_point: str = Field("standard", description="Операционная точка Speechmatics: standard или enhanced")
-    
-    @validator('speechmatics_operating_point')
-    def validate_speechmatics_operating_point(cls, v):
-        """Валидация operating_point для Speechmatics"""
-        valid_points = ['standard', 'enhanced']
-        if v not in valid_points:
-            raise ValueError(f"speechmatics_operating_point должен быть одним из: {', '.join(valid_points)}")
-        return v
-    
-    # Deepgram
-    deepgram_api_key: Optional[str] = Field(None, description="API ключ Deepgram для транскрипции и диаризации")
-    deepgram_model: str = Field("nova-2", description="Модель Deepgram для транскрипции (nova-2, whisper-cloud, enhanced, base)")
-    deepgram_language: str = Field("ru", description="Язык для Deepgram API")
-    
-    # Диаризация
-    enable_diarization: bool = Field(False, description="Включить диаризацию (разделение говорящих). По умолчанию отключено для стабильности.")
-    diarization_provider: str = Field("whisperx", description="Провайдер диаризации: whisperx, pyannote, picovoice")
-    huggingface_token: Optional[str] = Field(None, description="Токен Hugging Face для моделей диаризации")
-    picovoice_access_key: Optional[str] = Field(None, description="Access Key для Picovoice API")
-    leopard_model_path: Optional[str] = Field(None, description="Путь к .pv модели Picovoice Leopard (для выбора языка, например русский)")
-    diarization_device: str = Field("auto", description="Устройство для диаризации: auto, cpu, mps, cuda")
-    compute_type: str = Field("auto", description="Тип вычислений: auto, int8, float16, float32")
-    max_speakers: int = Field(10, description="Максимальное количество говорящих")
-    min_speakers: int = Field(1, description="Минимальное количество говорящих")
-    
-    # Очистка файлов
-    enable_cleanup: bool = Field(True, description="Включить автоматическую очистку временных файлов")
-    cleanup_interval_minutes: int = Field(30, description="Интервал очистки файлов в минутах")
-    temp_file_max_age_hours: int = Field(2, description="Максимальный возраст временных файлов в часах")
-    cache_max_age_hours: int = Field(24, description="Максимальный возраст кэш файлов в часах")
-    
-    # Улучшения качества протоколов
-    enable_diarization_analysis: bool = Field(True, description="Включить расширенный анализ данных диаризации")
-    enable_text_preprocessing: bool = Field(True, description="Включить предобработку текста транскрипции")
-    enable_protocol_validation: bool = Field(True, description="Включить валидацию и оценку качества протоколов")
-    meeting_type_detection: bool = Field(True, description="Включить автоопределение типа встречи")
-    
-      
-    # Оптимизация LLM пайплайна
-    enable_prompt_caching: bool = Field(True, description="Использовать prompt caching для OpenAI/Anthropic (экономия токенов)")
-    log_cache_metrics: bool = Field(True, description="Логировать детальные метрики кеширования токенов")
-    cache_metrics_in_final_message: bool = Field(False, description="Показывать метрики кеширования в финальном сообщении пользователю (для отладки)")
-    min_transcription_length_for_cache: int = Field(3000, description="Минимальная длина транскрипции (в символах) для активации кеширования. ~1024 токенов = ~3000-4000 символов")
-    max_context_tokens_stage2: int = Field(10000, description="Максимальное количество токенов контекста для Stage 2 (используются релевантные фрагменты)")
-
-    # Новая консолидированная архитектура (2 запроса вместо 5-6)
-    enable_consolidated_two_request: bool = Field(True, description="Включить новую консолидированную архитектуру (2 запроса вместо 5-6)")
-    consolidated_protocol_quality_threshold: float = Field(0.7, description="Порог качества для использования консолидированного метода (0.0-1.0)")
-    consolidated_max_tokens_request1: int = Field(4000, description="Максимальное токенов для запроса 1 (извлечение)")
-    consolidated_max_tokens_request2: int = Field(3000, description="Максимальное токенов для запроса 2 (протокол)")
-    consolidated_temperature_request1: float = Field(0.3, description="Температура для запроса 1 (высокая точность)")
-    consolidated_temperature_request2: float = Field(0.4, description="Температура для запроса 2 (качественное форматирование)")
-    
-    # Сопоставление участников (speaker mapping)
-    enable_speaker_mapping: bool = Field(True, description="Включить автоматическое сопоставление спикеров с участниками")
-    speaker_mapping_confidence_threshold: float = Field(0.7, description="Порог уверенности для сопоставления спикеров (0.0-1.0)")
-    max_participants: int = Field(20, description="Максимальное количество участников в списке")
-    llm_debug_log: bool = Field(False, description="Выводить в DEBUG лог все запросы и ответы LLM (генерация протоколов и сопоставление спикеров)")
-    include_raw_transcription_in_prompts: bool = Field(False, description="Включать исходную транскрипцию в промпты вместе с форматированной (для отладки проблем с диаризацией, увеличивает расход токенов)")
-    full_text_matching: bool = Field(False, description="Использовать полный текст транскрипции для сопоставления участников (увеличивает расход токенов)")
-    enable_speaker_mapping_confirmation: bool = Field(False, description="Показывать UI для подтверждения сопоставления спикеров перед генерацией протокола")
-    
-    # Настройки очереди задач
-    max_concurrent_tasks: Optional[int] = Field(None, description="Максимальное количество одновременно обрабатываемых задач (по умолчанию рассчитывается по CPU/RAM)")
-    max_queue_size: int = Field(100, description="Максимальный размер очереди задач")
-    queue_update_interval: float = Field(2.0, description="Интервал проверки изменений позиции в очереди (в секундах)")
-    queue_cleanup_interval_hours: int = Field(24, description="Интервал очистки завершенных задач из очереди (в часах)")
-    
-    # Администраторы
-    admins: List[int] = Field(default_factory=list, description="Список Telegram ID администраторов (через запятую)")
-    
-    @field_validator('admins', mode='before')
-    @classmethod
-    def parse_admins(cls, v):
-        """Парсинг списка администраторов из строки или списка"""
-        if not v:
-            return []
-        # Если уже int - оборачиваем в список
-        if isinstance(v, int):
-            return [v]
-        if isinstance(v, list):
-            return [int(x) for x in v]
-        if isinstance(v, str):
-            # Парсим строку с ID через запятую
-            try:
-                return [int(x.strip()) for x in v.split(',') if x.strip()]
-            except ValueError:
-                from loguru import logger
-                logger.warning(f"Некорректный формат ADMINS: {v}. Используется пустой список.")
-                return []
-        return []
-    
-    @validator('openai_models', pre=True, always=True)
-    def ensure_openai_models(cls, v, values):
-        """Гарантируем наличие хотя бы одного пресета OpenAI.
-        Если OPENAI_MODELS не задан, используем одиночные OPENAI_MODEL/OPENAI_BASE_URL.
-        """
-        try:
-            # Если явно передан список пресетов
-            if isinstance(v, list) and len(v) > 0:
-                # Приведение к целевой схеме произойдёт автоматически
-                return v
-        except Exception:
-            pass
-
-        # Формируем одиночный дефолтный пресет из старых настроек
-        model = values.get('openai_model', 'gpt-3.5-turbo')
-        base_url = values.get('openai_base_url')
-        default_preset = {
-            "key": "default",
-            "name": f"OpenAI: {model}",
-            "model": model,
-            "base_url": base_url
-        }
-        return [default_preset]
-    
-    model_config = SettingsConfigDict(
-        env_file=".env",
-        env_file_encoding="utf-8",
-        extra='ignore'
-    )
-
-
-settings = Settings()
+__all__ = ["Settings", "OpenAIModelPreset", "settings"]
